+++ flink-yarn-tests/src/test/java/org/apache/flink/yarn/YARNHighAvailabilityITCase.java
@@ -137,7 +137,7 @@ public class YARNHighAvailabilityITCase extends YarnTestBase {
 		HighAvailabilityServices highAvailabilityServices = null;
 
 		try {
-			yarnCluster = flinkYarnClient.deploySession();
+			yarnCluster = flinkYarnClient.deploySessionCluster();
 
 			final ClusterClient finalYarnCluster = yarnCluster;
