+++ flink-yarn/src/test/java/org/apache/flink/yarn/YarnClusterDescriptorTest.java
@@ -69,7 +69,7 @@ public class YarnClusterDescriptorTest {
 		clusterDescriptor.setTaskManagerSlots(Integer.MAX_VALUE);
 
 		try {
-			clusterDescriptor.deploySession();
+			clusterDescriptor.deploySessionCluster();
 
 			fail("The deploy call should have failed.");
 		} catch (RuntimeException e) {
@@ -98,7 +98,7 @@ public class YarnClusterDescriptorTest {
 		clusterDescriptor.setTaskManagerSlots(1);
 
 		try {
-			clusterDescriptor.deploySession();
+			clusterDescriptor.deploySessionCluster();
 
 			fail("The deploy call should have failed.");
 		} catch (RuntimeException e) {
